Refactoring and cleanup
