+++ src/test/groovy/main/MikeAndConquerTestBase.groovy
@@ -1,0 +1,33 @@
+package main
+
+import client.MikeAndConquerGameClient
+import spock.lang.Specification
+
+import javax.imageio.ImageIO
+import java.awt.image.BufferedImage
+
+
+class MikeAndConquerTestBase extends Specification {
+
+    MikeAndConquerGameClient gameClient
+
+    def setup() {
+        String localhost = "localhost"
+        String remoteHost = "192.168.0.146"
+//        String host = localhost
+        String host = remoteHost
+
+        int port = 11369
+        boolean useTimeouts = true
+//        boolean useTimeouts = false
+        gameClient = new MikeAndConquerGameClient(host, port, useTimeouts )
+        gameClient.resetGame()
+        gameClient.leftClickInWorldCoordinates(1,1)  // to get mouse clicks in default state
+    }
+
+
+
+
+
+
+}
